backend/tools_deployment/alzheimers-cloudrun/requirements.txt
flask
gunicorn
torch
torchvision
transformers
pillow
safetensors